refactor: split formatter_test.go into per-formatter test files
Split the 2040-line formatter_test.go into focused test files by
formatter type, consolidate repetitive tests into table-driven
patterns, and remove custom string helpers in favor of strings.Contains.

New files: compact_formatter_test.go, brief_formatter_test.go,
github_formatter_test.go, gitlab_formatter_test.go.
Color tests moved to color_test.go where they belong.
